# openapi.yaml
openapi: 3.1.0
info:
  title: FindTutor API
  version: 0.1.0
  description: >
    Contract for the FindTutor language-tutoring marketplace (Uzbekistan).
    Hand-written and treated as the source of truth: the Go backend renders
    responses to match it, and the Next.js frontend generates its typed client
    from it (openapi-typescript + openapi-fetch).

    Money is always an integer amount in MINOR units (tiyin for UZS, cents for
    USD) plus a currency code — never a float.

    Language: every human-readable `error.message` is rendered in the language
    of the request's `Accept-Language` header (`en`, `ru` or `uz`; anything
    else falls back to English) and the response carries `Content-Language`.
    Email is written in the account's stored `locale` (see AuthUser), set from
    Accept-Language at registration and changed via PATCH /v1/auth/me.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/teachers:
    get:
      tags: [teachers]
      summary: Search teachers
      operationId: listTeachers
      parameters:
        - name: language
          in: query
          description: Filter to teachers who teach this language (ISO 639-1 code, e.g. `en`).
          schema: { type: string }
        - name: kind
          in: query
          schema: { $ref: "#/components/schemas/TeacherKind" }
        - name: max_price_minor
          in: query
          description: Only teachers whose hourly price is at or below this amount (minor units).
          schema: { type: integer, format: int64, minimum: 0 }
        - name: q
          in: query
          description: Case-insensitive substring match across name, headline, and focus tags.
          schema: { type: string }
        - name: sort
          in: query
          schema:
            $ref: "#/components/schemas/TeacherSort"
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: A page of matching teachers plus catalog-wide facets.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherListResult" }
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [teachers]
      summary: Create the caller's teacher profile
      description: >
        Claims a teacher profile for the authenticated account. One profile per
        account — a second attempt is 409. The new profile's owner is the caller.

        The body carries only the editable fields; server-controlled aggregates
        (`rating`, `review_count`, `lessons_completed`, `student_count`,
        `response_time_hours`) start at 0, `accepting_students` starts true, and
        the `slug` is generated from `display_name` (de-duplicated with a numeric
        suffix). `languages` / `focus` / `experience` are stored as given.


        On create, `display_name`, `headline`, `kind`, `country_code`,
        `country_name`, `city`, `timezone`, and `price_per_hour_minor` are
        required; a missing or blank one is a 400.
      operationId: createTeacherProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeacherProfileWritable" }
      responses:
        "201":
          description: The created profile (same shape as GET /v1/teachers/{slug}).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherProfile" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The account already has a teacher profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/teachers/{slug}:
    get:
      tags: [teachers]
      summary: Get one teacher profile
      operationId: getTeacherBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The full teacher profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherProfile" }
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [teachers]
      summary: Edit the caller's own teacher profile
      description: >
        Partial update of the profile at `slug`. Requires a Bearer token whose
        account owns the profile (`teachers.user_id`); otherwise 401 (no/invalid
        token) or 403 (not the owner — an unclaimed profile is never editable).

        Every field is optional; an omitted field is left unchanged. A present
        `languages` / `focus` / `experience` array (even `[]`) fully replaces
        that collection. The `slug` is immutable and server-controlled
        aggregates cannot be set here.
      operationId: updateTeacherProfile
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeacherProfileWritable" }
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherProfile" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/teachers/{slug}/slots:
    get:
      tags: [bookings]
      summary: List a teacher's concrete bookable start times
      description: >
        Projects the teacher's recurring weekly availability
        (`teacher_availability_slots`, stored as UTC minutes-from-midnight on a
        UTC weekday) onto real UTC datetimes across `[from, to]`, then removes
        any start that collides with an existing non-cancelled booking.

        Window rules: `from` defaults to now (and is clamped forward to now),
        `to` defaults to `from + 14 days`, and `to - from` may not exceed 21
        days (400 otherwise). Slots are never generated in the past.

        A start time is included only if `[start, start + duration]` fits
        entirely inside one availability window. Windows the frontend split
        across 00:00 UTC are stitched back together first, so a lesson may
        legitimately straddle UTC midnight. Candidate starts step every 30
        minutes from the start of each window.

        `price.amount_minor = round(price_per_hour_minor * duration / 60)` in
        integer minor units, rounded half-up.
      operationId: listTeacherSlots
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: RFC3339 UTC. Defaults to now.
        - name: to
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: RFC3339 UTC. Defaults to from + 14 days. Max span 21 days.
        - name: duration
          in: query
          required: false
          schema: { type: integer, enum: [30, 60, 90, 120], default: 60 }
          description: Lesson length in minutes.
      responses:
        "200":
          description: The bookable start times in the window.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherSlots" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/bookings:
    post:
      tags: [bookings]
      summary: Book a lesson
      description: >
        Creates a booking in `pending_payment` and opens a `requires_payment`
        payment intent for its price. The server re-derives the bookable slot
        set and rejects a `start_at` that is not a real, currently free slot; it
        never trusts a client-supplied price.

        `duration_minutes` must be one of 30/60/90/120. When `is_trial` is true
        the duration is forced to 30 minutes and the price is the teacher's
        `trial_price_minor` (400 if the teacher offers no trial). You cannot
        book your own teacher profile.

        The student then calls `POST /v1/bookings/{id}/pay` to authorise
        payment, which confirms the booking.
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookingRequest" }
      responses:
        "201":
          description: The created booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: >
            The slot is not bookable (`slot_unavailable` — outside availability,
            misaligned, in the past, or already taken) or was taken by a
            concurrent request between the check and the insert (`slot_taken`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [bookings]
      summary: List the caller's bookings
      description: >
        With no `role`, returns every booking the caller participates in (as
        student or as the owner of the teacher profile). `role=student` limits
        it to bookings the caller booked; `role=teacher` to bookings for the
        teacher profile the caller owns (empty if they own none). Newest lesson
        first. Optional `status` filter.
      operationId: listBookings
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema: { type: string, enum: [student, teacher] }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/BookingStatus" }
      responses:
        "200":
          description: The matching bookings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingList" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/bookings/{id}:
    get:
      tags: [bookings]
      summary: Get one booking
      operationId: getBooking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/bookings/{id}/pay:
    post:
      tags: [bookings]
      summary: Pay for a booking
      description: >
        Authorises payment for a `pending_payment` booking and, on success,
        moves it to `confirmed` (the payment webhook performs the transition).
        Student-only — only the account that booked the lesson may pay.

        The `method_token` is an opaque payment-method handle. The MVP ships a
        deterministic fake provider (Stripe does not operate in Uzbekistan);
        its tokens drive the outcome:
        `pm_ok` authorises, `pm_decline` is refused (402 `payment_failed`),
        `pm_capture_fail` authorises but fails the first capture at completion.

        Response is the booking with its embedded `payment` (now `authorized`).
      operationId: payBooking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PayBookingRequest" }
      responses:
        "200":
          description: The confirmed booking with its authorised payment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not the student who booked this lesson (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "402":
          description: The payment method was declined (`payment_failed`). The booking stays `pending_payment`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            The booking is already paid (`already_paid`) or is cancelled
            (`invalid_state`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/bookings/{id}/complete:
    post:
      tags: [bookings]
      summary: Complete a lesson
      description: >
        Moves a `confirmed` booking to `completed` and captures the payment,
        then writes the teacher's payout-ledger entry. Teacher-owner only.
        Allowed only once the lesson's `end_at` is in the past (409 `too_early`
        otherwise).

        Capture happens before the status flips, so a capture failure leaves
        the booking `confirmed` and retryable (502 `capture_failed`).
      operationId: completeBooking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The completed booking with its captured payment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not the teacher for this lesson (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            The booking is not `confirmed` (`invalid_state`) or the lesson has
            not ended yet (`too_early`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: The provider could not capture the hold (`capture_failed`). The booking stays `confirmed`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/bookings/{id}/cancel:
    post:
      tags: [bookings]
      summary: Cancel a booking
      description: >
        Moves `pending_payment` or `confirmed` -> `cancelled`, recording
        `cancelled_at` and the optional `reason`. Participant-only. For the MVP
        cancellation is allowed at any time (no window / penalties yet). Already
        `completed` or `cancelled` -> 409.

        Payment side effects: an `authorized` intent is refunded in full (the
        hold is released); an intent that was never authorised is voided. If a
        payout-ledger row exists it is reversed.
      operationId: cancelBooking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CancelBookingRequest" }
      responses:
        "200":
          description: The cancelled booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The booking is already completed or cancelled (`invalid_state`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/bookings/{id}/meeting-link:
    put:
      tags: [bookings]
      summary: Set or clear a booking's meeting link
      description: >
        Sets the per-booking meeting-link override. Teacher-owner only. A
        non-empty `url` must be an http(s) URL; an empty string clears the
        override so the booking falls back to the teacher's default
        `meeting_url`.

        The link is still only revealed on the booking (`Booking.meeting_url`)
        once the booking is `confirmed` / `completed` and to a participant.
      operationId: setBookingMeetingLink
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MeetingLinkRequest" }
      responses:
        "200":
          description: The booking with the override applied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not the teacher for this lesson (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/bookings/{id}/no-show:
    post:
      tags: [bookings]
      summary: Record a no-show
      description: >
        Teacher-owner only (MVP). Allowed only from `confirmed` and only once
        `start_at` is in the past (409 `too_early` otherwise — a lesson has not
        been missed until it has started).

        `party = "student"` — the teacher attended, the student did not: the
        booking moves to `completed`, `no_show_party = student`, the payment is
        captured and the payout-ledger row written (identical to `complete`).

        `party = "teacher"` — the teacher self-reports they could not attend:
        the booking moves to `cancelled`, `no_show_party = teacher`, and the
        student is refunded (identical to `cancel`).
      operationId: recordBookingNoShow
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoShowRequest" }
      responses:
        "200":
          description: The updated booking (completed or cancelled).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not the teacher for this lesson (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            The booking is not `confirmed` (`invalid_state`) or the lesson has
            not started yet (`too_early`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/admin/bookings:
    get:
      tags: [admin]
      summary: Every booking on the platform
      description: >
        Permission: `bookings.view`. Unlike `GET /v1/bookings` (the caller's own
        lessons) this lists every booking, newest lesson first. `status` is an
        optional exact filter; `q` matches the teacher's display name or slug and
        the student's email or display name, case-insensitively.
      operationId: adminListBookings
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/BookingStatus" } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: A page of bookings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminBookingList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived HS256 access token from /v1/auth/{login,register,refresh}.
    cookieAuth:
      type: apiKey
      in: cookie
      name: ftr_session
      description: >
        HttpOnly opaque refresh token, Path=/v1/auth, SameSite=Lax. Set by the
        auth endpoints; the browser sends it automatically to /v1/auth/refresh
        and /v1/auth/logout. Not readable or settable by client JS.
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    TeacherKind:
      type: string
      enum: [professional, community]
      description: >
        `professional` — certified/qualified teacher, structured lessons.
        `community` — conversation practice, informal tutoring.
    TeacherSort:
      type: string
      enum: [recommended, price_asc, price_desc, rating_desc]
      default: recommended
...
    TeacherListResult:
      type: object
      required: [teachers, total, facets]
      properties:
        teachers:
          type: array
          items: { $ref: "#/components/schemas/TeacherSummary" }
        total:
          type: integer
          minimum: 0
          description: Total matches across all pages.
        facets:
          type: object
          required: [languages]
          properties:
            languages:
              type: array
              items: { $ref: "#/components/schemas/LanguageFacet" }
...
    BookingStatus:
      type: string
      enum: [pending_payment, confirmed, completed, cancelled]
...